Refactor entity collection into abstract class
